requirements.txt
accelerate==1.11.0
aiohttp==3.9.5
allennlp==2.10.0
aqlm==1.1.7
av==16.0.1
binpacking==1.5.2
bitsandbytes==0.48.2
dacite==1.8.1
datasets==3.6.0
decord==0.6.0
deepseek_vl==0.0.1.dev0
deepspeed==0.18.2
e2b_code_interpreter==2.4.1
eetq==1.0.0
einops==0.8.1
evalscope==1.3.0
fastapi==0.123.0
gradio==6.0.1
hqq==0.2.8.post1
huggingface_hub==0.35.3
janus==2.0.0
jieba==0.42.1
latex2sympy2_extended==1.10.2
librosa==0.11.0
liger_kernel==0.6.4
lmdeploy==0.10.2
math_verify==0.8.0
matplotlib==3.8.2
modelscope==1.31.0
moviepy==2.2.1
neraug==0.1.1
nltk==3.8.1
numpy==2.3.5
openai==2.8.1
opencompass==0.5.1
optimum==1.21.2
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
psutil==5.9.7
pyreft==0.1.0
python-dotenv==1.2.1
q_galore_torch==1.0
qwen_omni_utils==0.0.8
qwen_vl_utils==0.0.14
Requests==2.32.5
rouge==1.0.1
safetensors==0.7.0
scikit_learn==1.7.2
scipy==1.16.3
sentence_transformers==5.1.2
soundfile==0.13.1
swanlab==0.7.2
sympy==1.13.1
tensorboard==2.15.1
tokenizer==3.5.5
torch==2.6.0
torch_xla==2.9.0
torchacc==0.0.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
trl==0.16.1
unsloth==2025.11.4
uvicorn==0.38.0
vllm==0.11.2
wandb==0.12.21
whisper==1.1.10
